Datagen/requirements.txt
azure-storage-blob~=12.17.0
azure-storage-file-datalake~=12.12.0
azure-common==1.1.28
azure-core==1.28.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
azure-common==1.1.28
azure-core==1.28.0
azure-cosmos==4.3.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confluent-kafka==2.3.0
dataclasses-json==0.6.3
distlib==0.3.6
distro==1.8.0
et-xmlfile==1.1.0
fastapi==0.104.1
filelock==3.12.0
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.345
langchain-core==0.0.9
langsmith==0.0.69
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.5
pandas==2.1.3
pip-upgrade==0.0.6
platformdirs==3.2.0
pur==7.1.0
pydantic==2.5.2
pydantic_core==2.14.5
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.24.0.post1
virtualenv==20.22.0
virtualenv-clone==0.5.7
watchfiles==0.21.0
websockets==12.0
yarl==1.9.3
